# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BrandPulse AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/signals/{id}": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "Signal detail",
        "operationId": "getSignal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Signal UUID (or a demo alias when DEMO_MODE=true).",
              "title": "Id"
            },
            "description": "Signal UUID (or a demo alias when DEMO_MODE=true)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalDetail"
                }
              }
            }
          },
          "404": {
            "description": "not_found: unknown or deleted ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signals/{id}/investigate": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Start or reuse an investigation (idempotent by default)",
        "description": "`202` when a new investigation is created. `200` with `reused: true` when a queued, running or completed investigation already exists and `force` is false.",
        "operationId": "investigateSignal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Signal UUID (or a demo alias when DEMO_MODE=true).",
              "title": "Id"
            },
            "description": "Signal UUID (or a demo alias when DEMO_MODE=true)."
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Start a new investigation even if a completed one exists.",
              "default": false,
              "title": "Force"
            },
            "description": "Start a new investigation even if a completed one exists."
          },
          {
            "name": "X-Access-Code",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only when new SerpApi calls are needed",
              "title": "X-Access-Code"
            },
            "description": "Only when new SerpApi calls are needed"
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigateSignalResponse"
                }
              }
            }
          },
          "200": {
            "description": "Existing investigation reused (`reused: true`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigateSignalResponse"
                }
              }
            }
          },
          "403": {
            "description": "live_access_required: new SerpApi calls needed; `X-Access-Code` missing or wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found: unknown or deleted ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "analysis_not_ready, analysis_failed, signal_not_investigable, investigation_in_progress or live_data_disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "daily_limit_reached, rate_limited (`Retry-After` set) or serpapi_quota_low.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Status, then full report",
        "operationId": "getInvestigation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Investigation UUID (or a demo alias when DEMO_MODE=true).",
              "title": "Id"
            },
            "description": "Investigation UUID (or a demo alias when DEMO_MODE=true)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found: unknown or deleted ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{id}/evidence": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Evidence explorer",
        "operationId": "listInvestigationEvidence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Investigation UUID (or a demo alias when DEMO_MODE=true).",
              "title": "Id"
            },
            "description": "Investigation UUID (or a demo alias when DEMO_MODE=true)."
          },
          {
            "name": "stance",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Stance"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stance"
            }
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEvidenceResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found: unknown or deleted ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: field validation failed (`details.fields`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented: route is a contract stub until its phase lands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrandRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/BrandRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "BrandRef"
      },
      "BrandRole": {
        "type": "string",
        "enum": [
          "target",
          "competitor",
          "suggested"
        ],
        "title": "BrandRole"
      },
...
      "CompetitorComparison": {
        "properties": {
          "aspect": {
            "type": "string",
            "title": "Aspect"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ComparisonRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "aspect",
          "rows"
        ],
        "title": "CompetitorComparison"
      },
...
      "ConfidenceLabel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ConfidenceLabel",
        "description": "Derived at read time from the 0-100 score (DATABASE.md section 3)."
      },
      "ConfidenceSource": {
        "type": "string",
        "enum": [
          "signal",
          "investigation"
        ],
        "title": "ConfidenceSource",
        "description": "API.md section 3.4: which number `confidence` currently reflects."
      },
      "CoverageSourceType": {
        "type": "string",
        "enum": [
          "web",
          "news",
          "youtube",
          "forum",
          "shopping",
          "trends"
        ],
        "title": "CoverageSourceType",
        "description": "`source_coverage` entries: the source types plus `trends` (API.md section 3.8)."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "EvidenceCounts": {
        "properties": {
          "supports": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Supports"
          },
          "contradicts": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Contradicts"
          },
          "neutral": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Neutral"
          }
        },
        "type": "object",
        "required": [
          "supports",
          "contradicts",
          "neutral"
        ],
        "title": "EvidenceCounts"
      },
      "EvidenceItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stance": {
            "$ref": "#/components/schemas/Stance"
          },
          "relevance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "rank": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rank"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stance",
          "relevance",
          "rank",
          "source"
        ],
        "title": "EvidenceItem"
      },
      "Finding": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "evidence_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Finding"
      },
      "GeneratedBy": {
        "type": "string",
        "enum": [
          "llm",
          "fallback"
        ],
        "title": "GeneratedBy",
        "description": "API.md section 3.8: `fallback` = deterministic templates because Groq was unavailable."
      },
...
      "ImpactLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ImpactLevel"
      },
      "InvestigateSignalResponse": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Investigation Id"
          },
          "status": {
            "$ref": "#/components/schemas/InvestigationStatus"
          },
          "reused": {
            "type": "boolean",
            "title": "Reused"
          },
          "poll_url": {
            "type": "string",
            "title": "Poll Url"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "status",
          "reused",
          "poll_url"
        ],
        "title": "InvestigateSignalResponse"
      },
      "InvestigationConfidence": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "label": {
            "$ref": "#/components/schemas/ConfidenceLabel"
          },
          "factors": {
            "$ref": "#/components/schemas/ConfidenceFactors"
          }
        },
        "type": "object",
        "required": [
          "score",
          "label",
          "factors"
        ],
        "title": "InvestigationConfidence"
      },
      "InvestigationRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/InvestigationStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "InvestigationRef"
      },
      "InvestigationReport": {
        "properties": {
          "generated_by": {
            "$ref": "#/components/schemas/GeneratedBy"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "confidence": {
            "$ref": "#/components/schemas/InvestigationConfidence"
          },
          "source_coverage": {
            "items": {
              "$ref": "#/components/schemas/SourceCoverage"
            },
            "type": "array",
            "title": "Source Coverage"
          },
          "search_interest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportSearchInterest"
              },
              {
                "type": "null"
              }
            ]
          },
          "competitor_comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompetitorComparison"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "generated_by",
          "summary",
          "scope",
          "confidence",
          "disclaimer"
        ],
        "title": "InvestigationReport"
      },
      "InvestigationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "signal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Signal Id"
          },
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "status": {
            "$ref": "#/components/schemas/InvestigationStatus"
          },
          "step": {
            "$ref": "#/components/schemas/InvestigationStep"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/InvestigationStepInfo"
            },
            "type": "array",
            "title": "Steps"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "signal_id",
          "analysis_id",
          "status",
          "step"
        ],
        "title": "InvestigationResponse",
        "description": "`getInvestigation`. `report` is null until `completed`; `error` is user-safe on `failed`."
      },
      "InvestigationStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "InvestigationStatus"
      },
      "InvestigationStep": {
        "type": "string",
        "enum": [
          "generating_queries",
          "collecting_evidence",
          "scoring_evidence",
          "comparing_competitors",
          "synthesizing",
          "recommending",
          "done"
        ],
        "title": "InvestigationStep"
      },
      "InvestigationStepInfo": {
        "properties": {
          "key": {
            "$ref": "#/components/schemas/InvestigationStep"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "state": {
            "$ref": "#/components/schemas/StepState"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "state"
        ],
        "title": "InvestigationStepInfo"
      },
...
      "ListEvidenceResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "counts": {
            "$ref": "#/components/schemas/EvidenceCounts"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "counts"
        ],
        "title": "ListEvidenceResponse"
      },
...
      "Priority": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Priority"
      },
      "Recommendation": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "evidence_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          }
        },
        "type": "object",
        "required": [
          "id",
          "priority",
          "title",
          "action",
          "rationale",
          "timeframe"
        ],
        "title": "Recommendation"
      },
      "ReportSearchInterest": {
        "properties": {
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Pct"
          }
        },
        "type": "object",
        "required": [
          "keyword"
        ],
        "title": "ReportSearchInterest"
      },
      "Scope": {
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/ScopeVerdict"
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "verdict",
          "explanation"
        ],
        "title": "Scope"
      },
      "ScopeVerdict": {
        "type": "string",
        "enum": [
          "brand_specific",
          "industry_wide",
          "inconclusive",
          "unknown"
        ],
        "title": "ScopeVerdict"
      },
      "ScoreComponents": {
        "properties": {
          "growth": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Growth"
          },
          "frequency": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Frequency"
          },
          "cross_source": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Cross Source"
          },
          "sentiment_impact": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Sentiment Impact"
          }
        },
        "type": "object",
        "required": [
          "growth",
          "frequency",
          "cross_source",
          "sentiment_impact"
        ],
        "title": "ScoreComponents"
      },
...
      "SignalDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/SignalKind"
          },
          "aspect": {
            "type": "string",
            "title": "Aspect"
          },
          "growth": {
            "type": "number",
            "title": "Growth"
          },
          "impact": {
            "$ref": "#/components/schemas/ImpactLevel"
          },
          "confidence": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "confidence_source": {
            "$ref": "#/components/schemas/ConfidenceSource"
          },
          "sources_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sources Count"
          },
          "source_types": {
            "items": {
              "$ref": "#/components/schemas/SourceType"
            },
            "type": "array",
            "title": "Source Types"
          },
          "current_share": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Current Share"
          },
          "baseline_share": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Baseline Share"
          },
          "current_n": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Current N"
          },
          "baseline_n": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Baseline N"
          },
          "trend_corroborated": {
            "type": "boolean",
            "title": "Trend Corroborated"
          },
          "status": {
            "$ref": "#/components/schemas/SignalStatus"
          },
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "brand": {
            "$ref": "#/components/schemas/BrandRef"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "score_components": {
            "$ref": "#/components/schemas/ScoreComponents"
          },
          "latest_investigation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "aspect",
          "growth",
          "impact",
          "confidence",
          "confidence_source",
          "sources_count",
          "source_types",
          "current_share",
          "baseline_share",
          "current_n",
          "baseline_n",
          "trend_corroborated",
          "status",
          "analysis_id",
          "brand",
          "score",
          "score_components"
        ],
        "title": "SignalDetail",
        "description": "`getSignal`. Follows the API.md 3.6 example: `brand` (not `brand_id`) and\n`latest_investigation` (not `investigation_id`)."
      },
      "SignalKind": {
        "type": "string",
        "enum": [
          "aspect_negative_spike"
        ],
        "title": "SignalKind"
      },
      "SignalStatus": {
        "type": "string",
        "enum": [
          "detected",
          "investigating",
          "investigated"
        ],
        "title": "SignalStatus"
      },
...
      "SourceCoverage": {
        "properties": {
          "source_type": {
            "$ref": "#/components/schemas/CoverageSourceType"
          },
          "supporting": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Supporting"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "supporting"
        ],
        "title": "SourceCoverage"
      },
      "SourceType": {
        "type": "string",
        "enum": [
          "web",
          "news",
          "youtube",
          "forum",
          "shopping"
        ],
        "title": "SourceType"
      },
      "Stance": {
        "type": "string",
        "enum": [
          "supports",
          "contradicts",
          "neutral"
        ],
        "title": "Stance"
      },
      "StepState": {
        "type": "string",
        "enum": [
          "done",
          "active",
          "pending"
        ],
        "title": "StepState",
        "description": "State of one investigation step (API.md section 3.8)."
      },
...
    }
  }
}